refactor: change requests page's structure
